+++ s1tiling/libs/otbwrappers/lia.py
@@ -201,6 +201,7 @@ class ProjectDEMToS2Tile(ExecutableStepFactory):
             gen_output_filename=TemplateOutputFilenameGenerator(fname_fmt),
             image_description="Warped DEM to S2 tile",
         )
+        self.__dem_info          = cfg.dem_info
         self.__out_spatial_res   = cfg.out_spatial_res
         self.__resampling_method = cfg.dem_warp_resampling_method
         self.__nb_threads        = cfg.nb_procs
@@ -226,6 +227,7 @@ class ProjectDEMToS2Tile(ExecutableStepFactory):
         imd = meta['image_metadata']
         imd['S2_TILE_CORRESPONDING_CODE'] = meta['tile_name']
         imd['SPATIAL_RESOLUTION']         = str(self.__out_spatial_res)
+        imd['DEM_INFO']                   = self.__dem_info
         imd['DEM_RESAMPLING_METHOD']      = self.__resampling_method
         imd['ORTHORECTIFIED']             = 'true'
         # TODO: Import DEM_LIST from input VRT image
@@ -486,6 +488,7 @@ class ComputeGroundAndSatPositionsOnDEMFromEOF(OTBStepFactory):
         )
         self.__cfg = cfg  # Will be used to access cached DEM intersecting S2 tile
         self.__nodata = nodata_XYZ(cfg)
+        self.__dem_info = cfg.dem_info
 
     def _update_filename_meta_post_hook(self, meta: Meta) -> None:
         """
@@ -557,6 +560,7 @@ class ComputeGroundAndSatPositionsOnDEMFromEOF(OTBStepFactory):
         assert 'image_metadata' in meta
         imd = meta['image_metadata']
         imd['POLARIZATION']             = ""  # Clear polarization information (makes no sense here)
+        imd['DEM_INFO']                 = self.__dem_info
         imd['DEM_LIST']                 = ', '.join(meta['dems'])
         imd['band.DirectionToScanDEM*'] = ''
         imd['band.Gain']                = ''
@@ -649,6 +653,7 @@ class ComputeGroundAndSatPositionsOnDEM(OTBStepFactory):
         )
         self.__cfg = cfg  # Will be used to access cached DEM intersecting S2 tile
         self.__nodata = nodata_XYZ(cfg)
+        self.__dem_info = cfg.dem_info
 
     @staticmethod
     def reduce_inputs(inputs: List[Meta]) -> List:
@@ -771,6 +776,7 @@ class ComputeGroundAndSatPositionsOnDEM(OTBStepFactory):
         assert 'image_metadata' in meta
         imd = meta['image_metadata']
         imd['POLARIZATION']             = ""  # Clear polarization information (makes no sense here)
+        imd['DEM_INFO']                 = self.__dem_info
         imd['DEM_LIST']                 = ', '.join(meta['dems'])
         imd['band.DirectionToScanDEM*'] = ''
         imd['band.Gain']                = ''
@@ -1445,6 +1451,7 @@ class SARDEMProjection(OTBStepFactory):
         self.__dem_db_filepath     = cfg.dem_db_filepath
         self.__dem_field_ids       = cfg.dem_field_ids
         self.__dem_main_field_id   = cfg.dem_main_field_id
+        self.__dem_info            = cfg.dem_info
 
     def _update_filename_meta_pre_hook(self, meta: Meta) -> Meta:
         """
@@ -1492,6 +1499,7 @@ class SARDEMProjection(OTBStepFactory):
         assert 'image_metadata' in meta
         imd = meta['image_metadata']
         imd['POLARIZATION'] = ""  # Clear polarization information (makes no sense here)
+        imd['DEM_INFO']     = self.__dem_info
         imd['DEM_LIST']     = ', '.join(meta['dems'])
 
     def add_image_metadata(self, meta: Meta, app) -> None:
@@ -1869,6 +1877,7 @@ class ConcatenateLIA(_ConcatenatorFactory):
             'LIA'     : extended_filename_lia_degree(cfg),
             'sin_LIA' : extended_filename_lia_sin(cfg),
         }
+        self.__dem_info = cfg.dem_info
 
     def _update_filename_meta_post_hook(self, meta: Meta) -> None:
         """
@@ -1885,6 +1894,7 @@ class ConcatenateLIA(_ConcatenatorFactory):
         """
         super().update_image_metadata(meta, all_inputs)
         imd = meta['image_metadata']
+        imd['DEM_INFO']  = self.__dem_info
         imd['DEM_LIST']  = ""  # Clear DEM_LIST information (a merge of 2 lists should be done actually)
 
     def complete_meta(self, meta: Meta, all_inputs: InputList) -> Meta: